fix: cleaned up the folder structure
