+++ functions/zapier/DaoCreated.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express'
-import { CHAIN_INFO } from '../src/configs/chains'
-import { work } from '../src/utils/work'
+import { CHAIN_INFO } from '../../src/configs/chains'
+import { work } from '../../src/utils/work'
 
 async function daoCreated(req: Request, res: Response) {
 	const { chain } = req.body
